+++ wlr/OGUR/OGUR/Collision/CoordVerifier.cs
@@ -7,12 +7,12 @@ namespace OGUR.Collision
 {
     internal static class CoordVerifier
     {
-        public static bool IsValid(double x, double y)
+        public static bool IsValid(float x, float y)
         {
             return (x > 0 && y > 0 && x < XnaManager.WindowWidth && y < XnaManager.WindowHeight);
         }
 
-        public static bool IsBlocked(double x, double y)
+        public static bool IsBlocked(float x, float y)
         {
             var target = new Point2(x + SpriteInfo.Width/2, y + SpriteInfo.Height/2);
             return
