+++ wlr/OGUR/OGUR/Collision/HitTest.cs
@@ -26,15 +26,15 @@ namespace OGUR.Collision
             return CalculateDistance(x1,x2,y1,y2) <= 2*(SpriteInfo.Radius*SpriteInfo.Radius);
         }
 
-        public static double GetDistance(GameplayObject source, GameplayObject target)
+        public static float GetDistance(GameplayObject source, GameplayObject target)
         {
             return CalculateDistance(source.GetPosition().X, target.GetPosition().X, source.GetPosition().Y,
                                      target.GetPosition().Y);
         }
 
-        private static double CalculateDistance(float x1,float x2, float y1, float y2)
+        private static float CalculateDistance(float x1,float x2, float y1, float y2)
         {
-            return (Math.Pow((x1 - x2), 2) + Math.Pow((y1 - y2), 2));
+            return (float)(Math.Pow((x1 - x2), 2) + Math.Pow((y1 - y2), 2));
         }
     }
 }
