+++ wlr/OGUR/OGUR/Strategies/AttackPlayers.cs
@@ -17,20 +17,20 @@ namespace OGUR.Strategies
             //Every player is dead
             if (null != opponent)
             {
-                decimal leftVelocity = (opponent.GetPosition().X < target.GetPosition().X)
+                float leftVelocity = (opponent.GetPosition().X < target.GetPosition().X)
                                            ? -target.GetInt(StatType.MOVE_SPEED)
                                            : 0;
-                decimal rightVelocity = (opponent.GetPosition().X > target.GetPosition().X)
+                float rightVelocity = (opponent.GetPosition().X > target.GetPosition().X)
                                             ? target.GetInt(StatType.MOVE_SPEED)
                                             : 0;
-                decimal xVel = leftVelocity + rightVelocity;
-                decimal downVelocity = (opponent.GetPosition().Y > target.GetPosition().Y)
+                float xVel = leftVelocity + rightVelocity;
+                float downVelocity = (opponent.GetPosition().Y > target.GetPosition().Y)
                                            ? target.GetInt(StatType.MOVE_SPEED)
                                            : 0;
-                decimal upVelocity = (opponent.GetPosition().Y < target.GetPosition().Y)
+                float upVelocity = (opponent.GetPosition().Y < target.GetPosition().Y)
                                          ? -target.GetInt(StatType.MOVE_SPEED)
                                          : 0;
-                decimal yVel = downVelocity + upVelocity;
+                float yVel = downVelocity + upVelocity;
                 target.MoveIfPossible((int) xVel, (int) yVel);
             }
         }
